Various fixes and content
Fixed the weird spacing made by the windows by adding an additional
"background" to fill the holes.

Terrain and a map.

Temporarily changed the color of the building tops so it doesn't blend
with one of the ground textures

+++ LTexture.h
@@ -17,7 +17,7 @@ class LTexture
         void free();
 
         //Renders texture at given point
-        void render( int x, int y, SDL_Rect* clip, float scalex, float scaley);
+        void render( float x, float y, SDL_Rect* clip, float scalex, float scaley);
 
         //Gets image dimensions
         int getWidth();
@@ -28,8 +28,8 @@ class LTexture
         SDL_Texture* mTexture;
 
         //Image dimensions
-        int mWidth;
-        int mHeight;
+        float mWidth;
+        float mHeight;
 };
 
 void LTexture::free()
